sweet alert added to delete in admin view

+++ src/components/FeedbackItem/FeedbackItem.jsx
@@ -1,20 +1,38 @@
 import { TableRow, TableCell } from '@mui/material';
 import DeleteIcon from '@mui/icons-material/Delete';
 import axios from 'axios';
+import swal from 'sweetalert';
 
 function FeedbackItem({ feedback, fetchFeedback }) {
 
     const handleClick = () => {
-        console.log('inside handleClick');
-        axios.delete(`/feedback/${feedback.id}`)
-            .then((response) => {
-                console.log('Successful delete');
-                fetchFeedback();
-            })
-            .catch((err) => {
-                console.log('Error in delete', err);
-            })
+
+        swal({
+            title: "Are you sure?",
+            text: "Once deleted, you will not be able to recover this data!",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+        })
+            .then((willDelete) => {
+                if (willDelete) {
+                    axios.delete(`/feedback/${feedback.id}`)
+                        .then((response) => {
+                            console.log('Successful delete');
+                            fetchFeedback();
+                        })
+                        .catch((err) => {
+                            console.log('Error in delete', err);
+                        })
+                    swal("Feedback has been deleted!", {
+                        icon: "success",
+                    });
+                } else {
+                    swal("Your feedback is safe!");
+                }
+            });
     }
+
     return (
         <>
             <TableRow>
